+++ mcdc/Source/quinemccluskey.cpp
@@ -421,12 +421,12 @@ void QuineMcluskey::compareFromBitCountUpperToBitCountLower(uint indexReductionT
 	for (uint ui = indexUpper; ui > indexLower; --ui)
 	{
 		compareTwoEntries(	reductionTable[indexReductionTableColumn][ui],								// Subtable with set bit count x
-							reductionTable[indexReductionTableColumn][static_cast<ull>(ui) - 1],		// Subtable with set bit count x-1
+							reductionTable[indexReductionTableColumn][static_cast<uint>(ui) - 1],		// Subtable with set bit count x-1
 
 							// Write to other (next) reduction table at the lower bit count index
 							// It is alwyas the lower bit cout index, because if you comapre something
 							// with 4 bits sets and with 3 bits set, then the result is of youres something with 3 bits set
-							reductionTable[static_cast<ull>(indexReductionTableColumn) + 1][static_cast<ull>(ui) - 1]);
+							reductionTable[static_cast<uint>(indexReductionTableColumn) + 1][static_cast<uint>(ui) - 1]);
 	}
 }
